csl/moco/requirements.txt
# os, random, math, yaml, glob, pickle, copy, pathlib, cv2, sklearn
numpy==1.19.5
tqdm==4.41.1
pandas==1.1.5
matplotlib==3.2.2
Pillow==7.1.2
torch==1.8.1
torchvision==0.9.1
argparse==1.1
pytorch-metric-learning==0.9.98
faiss-gpu==1.7.0